Update tox environment (#744)

+++ gating/github_action_scripts/linting_run.sh
@@ -2,21 +2,23 @@
 
 set -x
 
-# Remove pip and virtualenv from container so we
-# can install our own.
-if which pip; then
-    sudo apt-get -y remove python-pip
-fi
-if which virtualenv; then
-    sudo apt-get -y remove virtualenv
-fi
+SELFINSTALL_PIP=false
 
-# Set to path to make sure we can hit the new pip and virtualenv commands
-export PATH=$PATH:/usr/local/bin:/usr/local/sbin:/home/runner/.local/bin:/home/runner/.local/sbin
+if [ $SELFINSTALL_PIP == true ]; then
+  if which pip; then
+      sudo apt-get -y remove python3-pip python-pip
+  fi
+  if which virtualenv; then
+      sudo apt-get -y remove virtualenv
+  fi
 
-# Download and run the pip installer
-curl --silent --show-error --retry 5 https://bootstrap.pypa.io/get-pip.py > /tmp/get-pip.py
-python2.7 /tmp/get-pip.py 'setuptools==39.0.1' 'virtualenv<20'
+  # Set to path to make sure we can hit the new pip and virtualenv commands
+  export PATH=$PATH:/usr/local/bin:/usr/local/sbin:/home/runner/.local/bin:/home/runner/.local/sbin
+
+  # Download and run the pip installer
+  curl --silent --show-error --retry 5 https://bootstrap.pypa.io/get-pip.py > /tmp/get-pip.py
+  python3 /tmp/get-pip.py 'setuptools==50.3.2' 'virtualenv==20.2.1'
+fi
 
 # Install some dependancies
 pip install bindep tox
